+++ packages/moneystream-extension/src/background/services/BackgroundScript.ts
@@ -31,12 +31,14 @@ import { LocalStorageProxy, STORAGE_KEY } from '../../types/storage'
 import { TabState } from '../../types/TabState'
 import { getFrameSpec, getTab } from '../../util/tabs'
 import { FrameSpec } from '../../types/FrameSpec'
-import { timeout } from '../../content/util/timeout'
+//import { timeout } from '../../content/util/timeout'
+import { Container } from 'inversify'
 
 import { StreamMoneyEvent } from './Stream'
 import { AuthService } from './AuthService'
 import { TabStates } from './TabStates'
 import { Streams } from './Streams'
+import { Offers } from './Offers'
 import { Favicons } from './Favicons'
 import { PopupBrowserAction } from './PopupBrowserAction'
 import { Logger, logger } from './utils'
@@ -45,12 +47,14 @@ import { MetanetService } from './MetanetService'
 import { BackgroundFramesService } from './BackgroundFramesService'
 import { StreamAssociations } from './StreamAssociations'
 import { Wallet } from 'moneystream-wallet'
+import { BandwidthTiers } from '@moneystream/polyfill-utils'
 
-// triggers monetized icon, was 1500
-const MONETIZED_TRIGGER = 650
+// triggers monetized icon
+const MONETIZED_TRIGGER = 600
 
 @injectable()
 export class BackgroundScript {
+  // private _tiers: BandwidthTiers
   constructor(
     private wallet: Wallet,
     private popup: PopupBrowserAction,
@@ -73,7 +77,8 @@ export class BackgroundScript {
     @inject(tokens.WextApi)
     private api: typeof window.chrome,
     @inject(tokens.MoneystreamDomain)
-    private moneystreamDomain: string
+    private moneystreamDomain: string,
+    private _offers: Offers
   ) {}
 
   get activeTab() {
@@ -90,7 +95,8 @@ export class BackgroundScript {
     this.tabStates.activeTab = value
   }
 
-  async run() {
+  async run(container: Container) {
+    //this._tiers = container.get(BandwidthTiers)
     this.initializeActiveTab()
     this.setRuntimeMessageListener()
     this.setTabsOnActivatedListener()
@@ -350,9 +356,9 @@ export class BackgroundScript {
   }
 
   // this makes envelope
-  sendResponseToContentScript(sender:MessageSender, responseData:any) {
+  sendResponseToContentScript(sender:MessageSender, command:string, responseData:any) {
     const { tabId, frameId } = getFrameSpec(sender)
-    const message = this.makePayloadEnvelope('info', responseData)
+    const message = this.makePayloadEnvelope(command, responseData)
     this.api.tabs.sendMessage(tabId, message, { frameId })
     return true
   }
@@ -369,7 +375,7 @@ export class BackgroundScript {
       case 'info':
         this.log('info command:', request.data)
         const response = await this.getInfoResponse()
-        sendResponse(this.sendResponseToContentScript(sender, response))
+        sendResponse(this.sendResponseToContentScript(sender, 'info', response))
         break
       case 'start':
           this.log('start command:', request.data)
@@ -446,8 +452,7 @@ export class BackgroundScript {
       case 'info':
         console.log(`Info BackgroundScript`)
         // this is required for browser
-        //const envelope = this.makePayloadEnvelope('info', await this.getInfoResponse())
-        sendResponse(this.sendResponseToContentScript(sender, await this.getInfoResponse()))
+        sendResponse(this.sendResponseToContentScript(sender, 'info', await this.getInfoResponse()))
         break
       case 'infodirect':
         console.log(`InfoDirect BackgroundScript`)
@@ -532,6 +537,12 @@ export class BackgroundScript {
       case 'onFrameAllowedChanged':
         sendResponse(await this.onFrameAllowedChanged(request, sender))
         break
+      case 'offer':
+        this._offers.add(request.data)
+        console.log(this._offers)
+        // TODO: process the offer, for now just respond ok
+        sendResponse(this.sendResponseToContentScript(sender, 'answer', "ok"))
+        break;
       default:
         sendResponse(false)
         break
@@ -711,8 +722,8 @@ export class BackgroundScript {
       this.store.playState = state.playState
       this.store.stickyState = state.stickyState
     } else if (state) {
-      delete this.store.playState
-      delete this.store.stickyState
+      this.store.playState = null
+      this.store.stickyState = null
     }
 
     if (state) {